+++ CHANGELOG.md
@@ -22,6 +22,9 @@
 - Upgrade to Worker SDK v14.4.0. [#1260](https://github.com/spatialos/gdk-for-unity/pull/1260)
   - The `Raknet`, `Tcp`, and `Kcp` network protocols have been deprecated. Please use `ModularKcp` and `ModularTcp` instead.
   - `ModularUdp` has been renamed to `ModularKcp`.
+- Changed the Worker SDK package installed for iOS. [#1268](https://github.com/spatialos/gdk-for-unity/pull/1268)
+    - Changed `c-static-fullylinked-arm-clang-ios` to `c-static-arm-clang-ios`.
+    - Changed `c-static-fullylinked-x86_64-clang-ios` to `c-static-x86_64-clang-ios`.
 
 ### Fixed